requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-openai
langchain-huggingface
azure-storage-blob
azure-identity
azure-search-documents
openpyxl